Adding modularizing
